LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/offset-rotate-expected.txt


PASS Can set 'offset-rotate' to CSS-wide keywords: initial
PASS Can set 'offset-rotate' to CSS-wide keywords: inherit
PASS Can set 'offset-rotate' to CSS-wide keywords: unset
PASS Can set 'offset-rotate' to CSS-wide keywords: revert
FAIL Can set 'offset-rotate' to var() references:  var(--A) assert_equals: expected 2 but got 1
FAIL Can set 'offset-rotate' to the 'auto' keyword: auto assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'offset-rotate' to the 'reverse' keyword: reverse assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'offset-rotate' to an angle: 0deg assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'offset-rotate' to an angle: 3.14rad assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'offset-rotate' to an angle: -3.14deg assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'offset-rotate' to an angle: calc(0rad + 0deg) assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSStyleValue]"
PASS Setting 'offset-rotate' to a length throws TypeError
PASS Setting 'offset-rotate' to a percent throws TypeError
PASS Setting 'offset-rotate' to a time throws TypeError
PASS Setting 'offset-rotate' to a flexible length throws TypeError
PASS Setting 'offset-rotate' to a number throws TypeError
PASS Setting 'offset-rotate' to a URL throws TypeError
PASS Setting 'offset-rotate' to a transform throws TypeError
PASS 'offset-rotate' does not supported 'auto 90deg'
PASS 'offset-rotate' does not supported 'reverse -90deg'
